// src/schema.gen.ts
export interface paths {
// ...
  "/samples": {
    /**
     * Sampleのリストを取得 GETリクエスト版 <br>
     *  検索条件については有無を気にしないでいいように、 Optional を用いて記述している。
     * @description Sampleのリストを取得 GETリクエスト版 <br>
     *  検索条件については有無を気にしないでいいように、 Optional を用いて記述している。
     */
    get: operations["findAllByGet"];
    /**
     * 新しいサンプルを作成 bodyからtitle等の値の入力を受け付ける。
     * @description 新しいサンプルを作成 bodyからtitle等の値の入力を受け付ける。
     */
    post: operations["create"];
  };
// ...
  "/samples/search": {
    /**
     * Sampleのリストを取得 POSTリクエスト版 <br>
     *  検索条件については有無を気にしないでいいように、 Optional を用いて記述している。
     * @description Sampleのリストを取得 POSTリクエスト版 <br>
     *  検索条件については有無を気にしないでいいように、 Optional を用いて記述している。
     */
    post: operations["findAllByPost"];
  };
// ...
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
// ...
    Error: {
      code: string;
      details: components["schemas"]["Detail"][];
      requestId: string;
      timestamp: string;
      userId: string;
    };
// ...
    SKIC06001E017Output: {
      /** @description 重複しているフォルダ名一覧 */
      duplicateFolderNames?: string[];
      /** @description 正常終了かどうか */
      ok?: boolean;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {

  /**
   * パスワード暗号化ツール
   *  パスワードの暗号化及び復号化を実施する
   *  (modekbn:1 [暗号化], 2[復号化])
   *  (strpassword: 暗号化/復号化を行う文字列)
   * @description パスワード暗号化ツール
   *  パスワードの暗号化及び復号化を実施する
   *  (modekbn:1 [暗号化], 2[復号化])
   *  (strpassword: 暗号化/復号化を行う文字列)
   */
  EncodePasswordTool: {
    parameters: {
      query: {
        mode: string;
        string: string;
      };
    };
// ...
  };
// ...
  createJob: {
    responses: {
      /** @description Created */
      201: {
        content: {
          "application/json": components["schemas"]["Job"];
        };
      };
      /** @description Unprocessable Entity */
      422: {
        content: {
          "*/*": components["schemas"]["Error"];
        };
      };
      /** @description Internal Server Error */
      500: {
        content: {
          "*/*": components["schemas"]["Error"];
        };
      };
    };
  };
// ...
}